# Aulas/Modulo_12/Swegger/docs/swagger.yaml
definitions:
  schemas.StudentResponse:
    properties:
      active:
        type: boolean
      age:
        type: integer
      cpf:
        type: integer
      createAt:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        description: DeletedAt time.Time `json:"deleteAt"`
        type: string
      updateAt:
        type: string
    type: object
info:
  contact: {}
paths:
  /students:
    get:
      consumes:
      - application/json
      description: Create student
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.StudentResponse'
        "400":
          description: Bad Request
      summary: Create student
      tags:
      - students
  /students/{id}:
    delete:
      consumes:
      - application/json
      description: Delete student details
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.StudentResponse'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete student
      tags:
      - students
    get:
      consumes:
      - application/json
      description: Retrieve students details using ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.StudentResponse'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get student by ID
      tags:
      - students
    put:
      consumes:
      - application/json
      description: Update student details
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.StudentResponse'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update student
      tags:
      - students
swagger: "2.0"